Shopicruit Task
This program was designed as part of Shopify's recruiting task. Given a
JSON file, the task was to determine how much the maximum computers and
keyboards that you could carry would cost you. There was also a
constraint on weight with a maximum of 100kg.

The main computation was written in JavaScript in conjunction with
jQuery so that the information is displayed on an HTML webpage.